+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- **dot**: show GPG passphrase prompt before pull decryption in CLI mode
+
 ## [0.6.1] - 2026-04-08
 
 ### Fixed